pyproject.toml
[project]
name = "customer-spending-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "mypy>=1.20.2",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "pymongo>=4.17.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
